Windows printf formats
git-svn-id: https://svn.r-project.org/R/trunk@64597 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/extra/tzone/strftime.c
@@ -275,8 +275,8 @@ static char *_fmt(const char *format, const stm *const t, char *pt, const char *
                 char buf[22]; // <= 19 digs + sign + terminator
                 int_fast64_t mkt = R_mktime(&tm);
 #ifdef WIN32
-                // will be limited in range for 32-bit long.
-                (void)snprintf(buf, 22, "%ld", (long)mkt);
+                // not ISO C99, so warns
+                (void)snprintf(buf, 22, "%I64d", mkt);
 #else
                 (void)snprintf(buf, 22, "%lld", (long long)mkt);
 #endif
